enums from constants.h moved into constants-fwd.h to be independent from any other headers.
